Fix a race condition in device tests
When running `cargo test --features device-test --test device` (with 14 threads
on my system), `test_last_device_removed_in_non_watch_mode` and
`test_wait_for_all_up_keys_up` sometimes fail:

test_wait_for_all_up_keys_up:

    stderr:
    Error: Device or resource busy (os error 16). Another program might have grabbed the device: 'test device eSubxmQCA6'

test_last_device_removed_in_non_watch_mode:

    stderr:
    Error: Device or resource busy (os error 16). Another program might have grabbed the device: 'test device dDCKpp5Suc'
    Error: Failed to prepare input devices: No device was selected!

This happens because `test_device_without_keys_is_not_selected_automatically`
performs automatic device selection and can grab virtual keyboards created by
other tests running in parallel.

Move the test back to `tests/device_2.rs`, which was removed in
https://github.com/xremap/xremap/pull/899/changes/681226b1d207c1425d9bc89ad948b5c54afed03a
(squashed as 4759ace1d9309d355ca0d1e9ab7e1e70ea1d9714), so Cargo runs it separately.
